% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{calculateCountDict}
\alias{calculateCountDict}
\title{calculateCountDict}
\usage{
calculateCountDict(count.vec)
}
\arguments{
\item{count.vec}{Interger vector, Selected count grid points}
}
\value{
count.dict
}
\description{
Calculate count dictionary from row value to approximated points
}
\details{
count dictionary is conversion vector from raw count to approximated count
}
\author{
Yasuhiro Kojima
}
