% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{calculateCountVec}
\alias{calculateCountVec}
\title{calculateCountVec}
\usage{
calculateCountVec(count.max, count.res, continuous.upper = 300)
}
\arguments{
\item{count.max}{Max values of \code{count.mat} in \code{link{Nbcd}}}

\item{count.res}{Approximation resolution of count}
}
\value{
count.vec Selected count grid points
}
\description{
Chose count values for approximation
}
\details{
Chose \code{count.res} count values from 0 ~ \code{max(count.mat)} for approximation. This chose all values up to min(continuos.upper, max(count.mat)) and chose other values based on bining of log(count).
}
\author{
Yasuhiro Kojima
}
