% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectSimulation.R
\name{calculateCtTransitionProb}
\alias{calculateCtTransitionProb}
\title{calculateCtTransitionProb}
\usage{
calculateCtTransitionProb(lqt, lpst, lambda)
}
\arguments{
\item{lqt}{Numeric vector, Each value represents Yt:n probability}

\item{lpst}{Numeric matrix, Log probability for each intervals}

\item{lambda}{Numeric vector, Each value represents Yt:n probability}
}
\value{
transition.prob.list List, each element represents vector of probability of next change point. index represents previous change points
}
\description{
Calculate transition probability for change points
}
\details{
This calculate the probability of next change points when we assume current change points are on each time points.
}
\author{
Yasuhiro Kojima
}
