% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChangeVariate.R
\name{calculateEntropyTwoCt}
\alias{calculateEntropyTwoCt}
\title{calculateEntropyTwoCt}
\usage{
calculateEntropyTwoCt(lqt, lpst, lambda)
}
\arguments{
\item{lqt}{Numeric vector, Each value represents Yt:n probability}

\item{lpst}{Numeric matrix, Log probability for each intervals}

\item{lambda}{Numeric vector, Each value represents Yt:n probability}
}
\value{
entropy Numeric, entropy of two change points
}
\description{
Calculate entropy of two change points
}
\details{
This is used for determin significant change variates
}
\author{
Yasuhiro Kojima
}
