% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLpst.R
\name{calculateLhrDivideTwo}
\alias{calculateLhrDivideTwo}
\title{calculatePst}
\usage{
calculateLhrDivideTwo(lpst, bin.prop = 0.2)
}
\arguments{
\item{lpst}{Numeric matrix, log probability for each time interval}
}
\value{
lhr.vec log ratio of compared probabilities
}
\description{
Calculate ratio of probability between divided and merged for local bin
}
\details{
Based on probability for each intervals, comparing probability for fixed length interval which is divided at half point or not.
}
\author{
Yasuhiro Kojima
}
