% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{calculateLnbValues}
\alias{calculateLnbValues}
\title{calculateLnbValues}
\usage{
calculateLnbValues(count.vec, p.vec, r)
}
\arguments{
\item{count.vec}{Interger vector, Selected count grid points}

\item{p.vec}{Interger vector, Selected p grid points}

\item{r}{Numeric, Size parameters of negative binomial distribution.}
}
\value{
lnb.mat
}
\description{
Calculate the probability value of log negative binomial distribution
}
\details{
The values are calculated for each count (row) and p (column)
}
\author{
Yasuhiro Kojima
}
