% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLpst.R
\name{calculateLpst}
\alias{calculateLpst}
\title{calculatePst}
\usage{
calculateLpst(lnb.dict, count.mat, t.grids)
}
\arguments{
\item{lnb.dict}{}

\item{count.mat}{}

\item{t.grids}{}
}
\value{
pst
}
\description{
Calculate the probability of count.mat[:, s:t]
}
\details{
Based on probability values of grid points in \code{lnb.dict}, this calculate the probability values of specific time intervals of  \code{count.mat}
}
\author{
小嶋泰弘
}
