% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLqt.R
\name{calculateLqt}
\alias{calculateLqt}
\title{Calculate Lqt}
\usage{
calculateLqt(lpst, lambda)
}
\arguments{
\item{lpst}{Numeric matrix, Log probability for each intervals}

\item{lambda}{Numeric, Prior probability for change point occurrence per step}
}
\value{
lqt Numeric vector, Each value represents Yt:n probability
}
\description{
Calculate Lqt, which represent probability of Yt:n given there is change point in t-1
}
\details{
These calculation are based on likelihood of each observation intervals
}
\author{
Yasuhiro Kojima
}
