% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMap.R
\name{calculateLr}
\alias{calculateLr}
\title{calculateLr}
\usage{
calculateLr(t, s, lql, lpst, lambda, time.num)
}
\arguments{
\item{t}{Integer, index of initial time points of unchange interval}

\item{s}{Integer, index of end time points of unchange interval}

\item{lql}{Numeric vector, log probability of interval t ~ time.num when change points are MAP}

\item{lpst}{Numeric matrix, log probability for each observation intervals}

\item{lambda}{Numeric, prior probability for change points in each transition}

\item{time.num}{Integer, index of end time points}
}
\value{
lr Numeric, statics for MAP estimate
}
\description{
Calculate statics for MAP estimate
}
\details{
Calculate the log probability assuming  there is no change points between t ~ s, and change points are MAP for s+1 ~ time.num
}
\author{
Yasuhiro Kojima
}
