% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMap.R
\name{calculateMap}
\alias{calculateMap}
\title{calculateMap}
\usage{
calculateMap(lpst, lambda)
}
\arguments{
\item{lpst}{Numeric matrix, log probability for each observation intervals}

\item{lambda}{Numeric, prior probability for change points in each transition}
}
\value{
map.change.point, Numeric vector, MAP estimates of change points
}
\description{
Calculation of map estimates for change points
}
\details{
This calculation is based on log probability for each observation intervals. see \code{\link{calculateLpst}}.
}
\author{
Yasuhiro Kojima
}
