% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNbcf.R
\name{calculateNbcf}
\alias{calculateNbcf}
\title{calculateNbcf}
\usage{
calculateNbcf(count.mat, t.vec, mean.bias.vec, alpha = 1, beta = 1,
  r = 30, lambda = 0.01, p.res = 1000, count.res = 1000,
  t.res = 100, map.fix.num = NULL, method = "nbinom")
}
\arguments{
\item{count.mat}{dgCmatrix, Count matrix. Each row represent each variate. Each column represent each observation}

\item{t.vec}{Numeric vector, Observation time}

\item{mean.bias.vec}{Numeric vector, Relative mean bias for each observation}

\item{r}{Numeric, Size parameters of negative binomial distribution.}

\item{lambda}{Numeric, Prior probability the change point occur each step}

\item{p.res}{Integer, Approximation resolution of parameter p}

\item{count.res}{Integer, Approximation resolution of count.}

\item{t.res}{Integer, Approximation resolution of t.}

\item{map.fix.num}{Integer, the number of map change points. If it is \code{NULL}, the point number is automaticaly estimated}

\item{method}{Character, this calculate least variance break points when you specify this as "variance"}
}
\value{
Nbcf, nbcf Instance of class \code{\link{Nbcf}}
}
\description{
Function for detecting change points and changed variates
}
\details{
This function is responsible for main calculation of this package. This calculate simulated change points and MAP estimates of them, and detect variates which significantly changed between each MAP estimated change points.
}
\seealso{
[Nbcf]
}
\author{
Yasuhiro Kojima
}
