% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLpst.R
\name{calculateOneVariateLpst}
\alias{calculateOneVariateLpst}
\title{calculateLpstOneVariate}
\usage{
calculateOneVariateLpst(lnb.dict, count.vec, t.grids)
}
\arguments{
\item{lnb.dict}{see \code{\link{lnb.dict}}}

\item{count.vec}{Integer vector, count data. each elements represent each observation}

\item{t.grids}{Numeric vector, grid index of t.vec}
}
\value{
lpst Numeric matrix, log probability for each time interval
}
\description{
Calculate the probability of count from  time poinst \code{s} to \code{t} for one variate
}
\details{
Based on probability values of grid points in \code{lnb.dict}, this calculate the probability values of specific time intervals of  \code{count.vec}. This returns a matrix whose row and column represent initial and end points of each interval
}
\author{
Yasuhiro Kojima
}
