% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{calculatePdict}
\alias{calculatePdict}
\title{calculatePvec}
\usage{
calculatePdict(p.vec, mean.bias.vec)
}
\arguments{
\item{p.res}{Integer, Approximation resolution of p}
}
\value{
p.dict Numeric matrix, conversion index form each grid point in each observation to reference grid points. each row represents grid point. each column represents each observation.
}
\description{
Calculate p dictionary from each observation grid to grids corrected for mean.bias.vec
}
\details{
We have to calculate negative binomial probability to p corrected for mean.bias.vec Hence, we need to use conversion before after correction.
}
\author{
Yasuhiro Kojima
}
