% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{calculatePvec}
\alias{calculatePvec}
\title{calculatePvec}
\usage{
calculatePvec(p.res)
}
\arguments{
\item{p.res}{Integer, Approximation resolution of p}
}
\value{
p.vec Numeric vector, Grid points for p
}
\description{
Calculation for grid points of p
}
\details{
These grids points are concentrated on 0 or 1. Use quadratic grids (Ferrer et al. 2016)
}
\author{
Yasuhiro Kojima
}
