% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{calculatePwidthVec}
\alias{calculatePwidthVec}
\title{calculatePwidthVec}
\usage{
calculatePwidthVec(p.vec)
}
\arguments{
\item{p.vec}{Numeric vector, Grid points for p}
}
\value{
p.width.vec grid width of p
}
\description{
Calculation for grid width of p
}
\details{
Calculation is based on \code{p.vec}
}
\author{
Yasuhiro Kojima
}
