% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLpst.R
\name{calculateVarLpstList}
\alias{calculateVarLpstList}
\title{calculateLpstVar}
\usage{
calculateVarLpstList(count.mat, mean.bias.vec, t.grids)
}
\arguments{
\item{count.mat}{Integer matrix, count data. each row and column represent variate and observation}

\item{mean.bias.vec}{Numeric vector, bias for each observation}

\item{t.grids}{Numeric vector, grid index of t.vec}
}
\value{
lpst.list List, variance of log normalized counts at each interval and each variate
}
\description{
Calculate variance of log normalized count for each intervals and each variate
}
\details{
Calculate variance of log normalized count for each intervals and each variate. This is used when \code{method == "variance"} in calculateNbcf.
}
\author{
Yasuhiro Kojima
}
