% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChangeVariate.R
\name{findChangeVariate}
\alias{findChangeVariate}
\title{findChangeVariate}
\usage{
findChangeVariate(lpst.list, used.vars, ct.vec, lambda = 0.5)
}
\arguments{
\item{lpst.list}{list of Numeric matrix, list of log probability for each intervals for each variate}

\item{used.vars}{vector, names of variates used in change point finding}

\item{ct.vec}{Numeric vector, change point vector}
}
\value{
change.var.df tibble, contains variate names, change points and bayes factors
}
\description{
Find change variates at each change points
}
\details{
This function calculate the signicance of  change points for each variate variates.
}
\author{
Yasuhiro Kojima
}
