% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectSimulation.R
\name{perfectSimulation}
\alias{perfectSimulation}
\title{perfectSimulation}
\usage{
perfectSimulation(lqt, lpst, lambda, sim.iter = 100)
}
\arguments{
\item{lqt}{Numeric vector, Each value represents Yt:n probability}

\item{lpst}{Numeric matrix, Log probability for each intervals}

\item{lambda}{Numeric vector, Each value represents Yt:n probability}

\item{sim.iter}{Integer, Iteration of simulation}
}
\value{
change.points.list List, list of numeric vector of change points at each simulation
}
\description{
Conduct perfectsimulation for change point
}
\details{
This function samples change points using perfectsimulation, which is a machine learning techunique.
}
\author{
Yasuhiro Kojima
}
