% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNbcf.R
\name{recalculateNbcf}
\alias{recalculateNbcf}
\title{recalculateNbcf}
\usage{
recalculateNbcf(old.nbcf, map.fix.num, lambda = 0.5, used.vars = NULL,
  calc.change.variate = TRUE)
}
\arguments{
\item{old.nbcf, }{nbcf Instance of class \code{\link{Nbcf}}}

\item{map.fix.num}{Integer, the number of map change points. If it is \code{NULL}, the point number is automaticaly estimated}

\item{method}{Character, this calculate least variance break points when you specify this as "variance"}
}
\value{
nbcf, Nbcf Instance of class \code{\link{Nbcf}}
}
\description{
Function for recalculation of part of \code{\link{calculateNbcf}}
}
\details{
This function recalculate \code{map.change.points} and \code{change.variate.df} from a existing instance of \code{nbcf}.
}
\seealso{
[Nbcf]
}
\author{
Yasuhiro Kojima
}
