% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated time course negative binomial data}
\format{An object of class \code{list} of length 2.}
\usage{
data(sim.data)
}
\description{
This data contains 6 variate time course negative binomial data.
elemnts of t.vec are 0.01, 0.02, ..., 1.0, and includes two change points,
where each variate can change the parameter p.
}
\examples{
sim.data <- data(sim.data)
nbcf.obj <- loadData(sim.data$count.mat, sim.data$t)
nbcf.obj <- setupNegativeBinomDict(nbcf.obj)
nbcf.obj <- calculateQt(nbcf.obj, lambda= 0.01, alpha=1.0, beta=1.0)
nbcf.obj <- perfectSimulation(nbcf.obj)
nbcf.obj <- estimateMap(nbcf.obj)
nbcf.obj <- detectVariate(nbcf.obj)
hist(nbcf.obj@change.point)
nbcf.obj@change.variate
}
\keyword{datasets}
