% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMap.R
\name{traceGetShatl}
\alias{traceGetShatl}
\title{traceGetShatl}
\usage{
traceGetShatl(shatl, taul)
}
\arguments{
\item{shatl}{Numeric vector, next MAP change points of each index, the end end elements should be -1}

\item{taul}{Numeric vector, previous MAP change points}
}
\value{
next.taul Numeric vector, one change points added for taul unless \code{shatl[taul[1] + 1] > 0}
}
\description{
Get trace of shatl
}
\details{
This recursively get the elements of MAP change points, like \code{next.tau = shatl[tau + 1]}
}
\author{
小嶋泰弘
}
