% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLpst.R
\name{calculateLpstList}
\alias{calculateLpstList}
\title{calculatePst}
\usage{
calculateLpstList(lnb.dict, count.mat, t.grids)
}
\arguments{
\item{lnb.dict}{see \code{\link{lnb.dict}}}

\item{count.mat}{Integer matrix, count data. each row and column represent variate and observation}

\item{t.grids}{Numeric vector, grid index of t.vec}
}
\value{
lpst Numeric matrix, log probability for each time interval
}
\description{
Calculate the probability of count.mat[:, s:t] variate-wise
}
\details{
Based on probability values of grid points in \code{lnb.dict}, this calculate the probability values of specific time intervals of  \code{count.mat}. This returns the list for each variate
}
\author{
Yasuhiro Kojima
}
