% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNbcf.R
\name{calculateNbcf}
\alias{calculateNbcf}
\title{calculateNbcf}
\usage{
calculateNbcf(count.mat, t.vec, mean.bias.vec, alpha = 1, beta = 1,
  r = 30, lambda = 0.01, p.res = 1000, count.res = 1000, t.res)
}
\arguments{
\item{count.mat}{Count matrix. Each row represent each variate. Each column represent each observation}

\item{t.vec}{Observation time}

\item{mean.bias.vec}{Relative mean bias for each observation}

\item{alpha}{Parameter fo beta distribution, which is prior for p of negative binomial distribution}

\item{beta}{Parameter fo beta distribution, which is prior for p of negative binomial distribution}

\item{r}{Size parameters of negative binomial distribution.}

\item{lambda}{Prior probability the change point occur each step}

\item{p.res}{Approximation resolution of parameter p}

\item{count.res}{Approximation resolution of count.}

\item{t.res}{Approximation resolution of t.}
}
\value{
nbcf Instance of class \code{\link{Nbcf}}
}
\description{
Function for detecting change points and changed variates
}
\details{
This function is responsible for main calculation of this package. This calculate simulated change points and MAP estimates of them, and detect variates which significantly changed between each MAP estimated change points.
}
\author{
Yasuhiro Kojima
}
