% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChangeVariate.R
\name{findChangeVariate}
\alias{findChangeVariate}
\title{findChangeVariate}
\usage{
findChangeVariate(lpst.list, used.vars, ct.vec, lambda = 0.5)
}
\arguments{
\item{lpst.list}{list of Numeric matrix, list of log probability for each intervals for each variate}

\item{used.vars}{vector, names of variates used in change point finding}

\item{ct.vec}{Numeric vector, change point vector}
}
\value{
change.var.df tibble, contains variate names, corresponding change point, entropy and change number among corresponding intervals
}
\description{
Find change variates at each change points
}
\details{
This function calculate the relation ships between change points and variates which have one change points among previous and following change points for each change point. Furthermore, we calculate the entropy of the change point and following change point.
}
\author{
Yasuhiro Kojima
}
