% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nbcf.R
\docType{class}
\name{Nbcf-class}
\alias{Nbcf-class}
\title{Data stroge class which contain the information required for finding change points.}
\description{
This class is a data storage object for calculating change points and changing variate.
You should firstly apply calculateQt and after either perfectSimulation or estimateMap.
}
\section{Slots}{

\describe{
\item{\code{count.mat}}{dgCMatrix, whose rows represent time, and columns represent variate.}

\item{\code{t.vec}}{Numeric vector, whose elements represent observed time corresponding to each row of count matrix.}

\item{\code{mean.bias.vec}}{Numeric vector, whose elements represent expted bias for mean value of each row.}

\item{\code{params}}{List, named list of parameters, whose mebers must be "lambda", "r", "alpha" and "beta".}

\item{\code{lnb.dict}}{LnbDict, Storage for precomputated log negative binomial probability. This is a instance of  \code{\link{LnbDict}}.}

\item{\code{Numeric}}{vector, Qt probability of Yt:n given there is change point in t-1}

\item{\code{Numeric}}{vector, sim.change.point change points simulated from perfectSimulation}

\item{\code{Numeric.vector,}}{map.change.point change points simulated from estimateMap}

\item{\code{Character}}{vector, change.variate named list of changed variates for each map.change.points}
}}

