% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcf.R
\docType{class}
\name{nbcf-class}
\alias{nbcf-class}
\alias{nbcf}
\title{Data stroge class which contain the information required for finding change points.}
\description{
This class is a data storage object for calculating change points and changing variate.
You should firstly apply calculateQt and after either perfectSimulation or estimateMap.
}
\section{Slots}{

\describe{
\item{\code{count.mat}}{Object of class \code{"ANY"} This is a countmatrix, whose rows represent time, and columns represent variate.}

\item{\code{t.vec}}{This is a numeric vector, whose elements represent observed time corresponding to each row of count matrix.}

\item{\code{params}}{This is a named list of parameters, whose mebers must be "lambda", "r", "alpha" and "beta".}

\item{\code{Qt}}{probability of Yt:n given there is change point in t-1}

\item{\code{nb.dict.mat}}{Storage for precomputated negative binomial probability.}

\item{\code{sim.change.point}}{change points simulated from perfectSimulation}

\item{\code{map.change.point}}{change points simulated from estimateMap}

\item{\code{change.variate}}{named list of changed variates for each map.change.points}
}}

