% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nbcf.R
\docType{class}
\name{Nbcf-class}
\alias{Nbcf-class}
\title{Data stroge class which contain the information required for finding change points.}
\description{
This class is a data storage object for calculating change points and changing variate.
You should firstly apply calculateQt and after either perfectSimulation or estimateMap.
}
\section{Slots}{

\describe{
\item{\code{count.mat}}{dgCMatrix, whose rows represent time, and columns represent variate.}

\item{\code{t.vec}}{Numeric vector, whose elements represent observed time corresponding to each row of count matrix.}

\item{\code{mean.bias.vec}}{Numeric vector, whose elements represent expted bias for mean value of each row.}

\item{\code{params}}{List, named list of parameters, whose mebers must be "lambda", "r", "alpha" and "beta".}

\item{\code{lpst}}{Numeric matrix, log probability of Ys:t}

\item{\code{lqt}}{Numeric vector, log probability of Yt:n given there is change point in t-1}

\item{\code{lpst.list}}{Numeric matrix, list of \code{lpst} for each variate}

\item{\code{lqt.list}}{Numeric matrix, list of \code{lqt} for each variate}

\item{\code{bf.list}}{Numeric matrix, list of bayes factor for each variate. The larger values indicate the variate changed in time course.}

\item{\code{used.vars}}{Character vector, variates used for change points calculation}

\item{\code{sim.change.point.list}}{List, each elements contains change points simulated from perfectSimulation}

\item{\code{map.change.point}}{Numeric.vector, change points simulated from estimateMap}

\item{\code{change.variate.df}}{tbl, data frame of changed variates. There are columns for change points, variate and etropy of change points. Entropy is expected to be lower for moredistinct change points.}
}}

