% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLnbDict.R
\name{setupLnbDict}
\alias{setupLnbDict}
\title{setupLnbDict}
\usage{
setupLnbDict(count.mat, p.res, count.res, mean.bias.vec)
}
\arguments{
\item{count.mat}{Count matrix. Each row represent each variate. Each column represent each observation}

\item{p.res}{Approximation resolution of parameter p}

\item{count.res}{Approximation resolution of count.}

\item{mean.bias.vec}{Relative mean bias for each observation}
}
\value{
lnbdict Instance of class \code{\link{LnbDict}}
}
\description{
Set up \code{\link{LnbDict}} instance
}
\details{
This function is responsible for generating instance of \code{\link{LnbDict}}. This involves calculate negative binomial probability for grid points of count and p. Futhermore, this calculate index conversion for count and p (see details in \code{\link{LnbDict}}.
}
\seealso{
[LnbDict]
}
\author{
Yasuhiro Kojima
}
